Door Update
Doors now can be set to constant state, meaning  they will remain open or closed when interacted with until they are interacted with again, NOT being reset by the fusebox reset.
Additionally the door meshes have been applied to the door blueprint so it now fits in the environment nicely.