Player Spawn now blocks navmesh
Player spawn collision block now prevents navmesh from being created inside it, meaning guards will not be able to walk into a spawn. If already placed, may have to be jiggled around to correctly update in the level. Pressing P will show the navmesh in the editor making it clear whether it is correctly positioned or not.